requirements.txt
google-adk
google-generativeai==0.4.0
gradio
litellm
python-dotenv
pdfplumber
sentence-transformers
faiss-cpu